refactor(input-label): change input-label for label
BREAKING CHANGE: input-label won't be available anymore

fix #49
