--- src/components/atoms/InputLabel/InputLabel.md
@@ -1,11 +1,0 @@
-### Summary
-
-Use `<InputLabel />` to render a native HTML `<label />`.
-
-### Example
-
-```jsx
-import { InputLabel } from '@zopauk/react-components';
-
-<InputLabel htmlFor="username">Username</InputLabel>;
-```
