+++ src/components/atoms/Label/Label.md
@@ -1,0 +1,11 @@
+### Summary
+
+Use `<Label />` to render a native HTML `<label />`.
+
+### Example
+
+```jsx
+import { Label } from '@zopauk/react-components';
+
+<Label htmlFor="username">Username</Label>;
+```
