+++ src/components/atoms/Label/Label.tsx
@@ -3,14 +3,14 @@ import styled from 'styled-components';
 import Text from '../Text/Text';
 import { colors } from '../../../constants/colors';
 
-export interface IInputLabelProps extends React.LabelHTMLAttributes<HTMLLabelElement> {}
+export interface ILabelProps extends React.LabelHTMLAttributes<HTMLLabelElement> {}
 
 const StyledLabel = styled(Text)`
   display: block;
 `;
 
-const InputLabel = (props: IInputLabelProps) => (
+const Label = (props: ILabelProps) => (
   <StyledLabel {...props} weight="semibold" forwardedAs="label" color={colors.neutral.dark} size="base" />
 );
 
-export default InputLabel;
+export default Label;
