+++ src/components/molecules/DropdownFiltered/DropdownFiltered.tsx
@@ -4,7 +4,7 @@ import styled, { css } from 'styled-components';
 import { colors } from '../../../constants/colors';
 import { typography } from '../../../constants/typography';
 import ErrorMessage from '../../atoms/ErrorMessage/ErrorMessage';
-import InputLabel from '../../atoms/InputLabel/InputLabel';
+import Label from '../../atoms/Label/Label';
 import InputText from '../../atoms/InputText/InputText';
 import Chevron from '../../icons/Chevron/Chevron';
 import SizedContainer from '../../layout/SizedContainer/SizedContainer';
@@ -151,9 +151,9 @@ export default class DropdownFiltered extends React.PureComponent<IDropdownFilte
     return (
       <div>
         {label && (
-          <InputLabel {...getLabelProps({ ...labelProps })} htmlFor={`text-id-${name}`}>
+          <Label {...getLabelProps({ ...labelProps })} htmlFor={`text-id-${name}`}>
             {label}
-          </InputLabel>
+          </Label>
         )}
         <SearchInputWrap>
           <SearchInput
